got rid of warning 'signed and unsigned type in conditional expression'
git-svn-id: https://svn.codehaus.org/groovy/trunk/groovy/modules/native_launcher@9389 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ source/jst_dynmem.c
@@ -55,7 +55,7 @@ extern void* jst_appendArrayItem( void* array, int indx, size_t* arlen, void* it
 
   // allocate array if requested
   if ( !array ) {
-    size_t size = *arlen > (size_t)indx + 1 ? *arlen : indx + 1 ;
+    size_t size = ( *arlen > (size_t)( indx + 1 ) ) ? *arlen : (size_t)( indx + 1 ) ;
     if ( ! ( array = jst_calloc( size, item_size_in_bytes ) ) ) return NULL ;
   } else if ( ((size_t)indx) >= *arlen ) { // ensure there is enough space
     size_t previousSize = *arlen ;
